# specs/api/rest/pokemon.openapi.yaml
openapi: 3.0.0

info:
  title: Pokemon API
  version: 1.0.0
  description: REST API for managing Pokemons

servers:
  - url: /api/v1

paths:
  /pokemons:
    get:
      summary: List all Pokemons with optional filters, pagination, and sorting
      parameters:
        - name: type
          in: query
          description: Filter by Pokemon type (exact match)
          required: false
          schema:
            type: string
          example: Electric
        - name: name
          in: query
          description: Filter by partial name (case-insensitive)
          required: false
          schema:
            type: string
          example: pika
        - name: sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            default: name
            enum: [name]
        - name: sortOrder
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            default: asc
            enum: [asc, desc]
        - name: page
          in: query
          description: Page number (starts at 1)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: A paginated list of Pokemons
          headers:
            X-RateLimit-Limit:
              description: Maximum requests allowed per minute
              schema:
                type: integer
              example: 100
            X-RateLimit-Remaining:
              description: Remaining requests in current window
              schema:
                type: integer
              example: 87
            X-RateLimit-Reset:
              description: Unix timestamp when rate limit resets
              schema:
                type: integer
              example: 1674567890
            Cache-Control:
              description: Cache directives
              schema:
                type: string
              example: public, max-age=300
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PokemonListResponse'
        '400':
          description: Invalid query parameters
        '429':
          description: Rate limit exceeded
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
            Retry-After:
              description: Seconds to wait before retrying
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'

    post:
      summary: Create a new Pokemon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePokemonInput'
      responses:
        '201':
          description: Pokemon created successfully
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pokemon'
        '400':
          description: Invalid input data
        '409':
          description: Pokemon with the same name already exists
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'

  /pokemons/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer

    patch:
      summary: Update an existing Pokemon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePokemonInput'
      responses:
        '200':
          description: Pokemon updated successfully
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pokemon'
        '400':
          description: Invalid input data
        '404':
          description: Pokemon not found
        '409':
          description: Pokemon name already exists
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'

    delete:
      summary: Delete a Pokemon
      responses:
        '204':
          description: Pokemon deleted successfully
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
        '404':
          description: Pokemon not found
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'

  /pokemons/import:
    post:
      summary: Import a Pokemon triggers a fetch from PokeAPI and adds/updates the record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportPokemonInput'
      responses:
        '200':
          description: Pokemon imported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pokemon'
        '400':
          description: Invalid input (e.g. non-positive ID)
        '404':
          description: Pokemon not found in PokeAPI
        '502':
          description: Bad Gateway - Failed to fetch from PokeAPI
        '504':
          description: Gateway Timeout - PokeAPI request timed out

components:
  schemas:
    Type:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - created_at

    Pokemon:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        types:
          type: array
          items:
            $ref: '#/components/schemas/Type'
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - types
        - created_at

    CreatePokemonInput:
      type: object
      properties:
        id:
          type: integer
          description: Client-provided Pokemon ID
        name:
          type: string
        types:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - types

    UpdatePokemonInput:
      type: object
      properties:
        name:
          type: string
        types:
          type: array
          items:
            type: string
      additionalProperties: false

    ImportPokemonInput:
      type: object
      properties:
        id:
          type: integer
          description: The ID of the Pokemon to import from PokeAPI
          example: 158
      required:
        - id

    PokemonListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Pokemon'
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'
      required:
        - data
        - pagination

    PaginationMetadata:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
          example: 1
        limit:
          type: integer
          description: Items per page
          example: 20
        totalCount:
          type: integer
          description: Total number of items
          example: 150
        totalPages:
          type: integer
          description: Total number of pages
          example: 8
      required:
        - page
        - limit
        - totalCount
        - totalPages

    RateLimitError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 429
        message:
          type: string
          example: Rate limit exceeded. Please try again later.
        error:
          type: string
          example: Too Many Requests
      required:
        - statusCode
        - message
        - error